+++ lib/terradactyl/cli.rb
@@ -1,5 +1,24 @@
 # frozen_string_literal: true
 
+# Fix for https://github.com/erikhuda/thor/issues/398
+class Thor
+  module Shell
+    class Basic
+      def print_wrapped(message, options = {})
+        indent = (options[:indent] || 0).to_i
+        if indent.zero?
+          stdout.puts message
+        else
+          message.each_line do |message_line|
+            stdout.print ' ' * indent
+            stdout.puts message_line.chomp
+          end
+        end
+      end
+    end
+  end
+end
+
 module Terradactyl
   # rubocop:disable Metrics/ClassLength
   class CLI < Thor
@@ -14,6 +33,7 @@ module Terradactyl
       super
     end
 
+    # rubocop:disable Metrics/BlockLength
     no_commands do
       # Monkey-patch Thor internal method to break out of nested calls
       def invoke_command(command, *args)
@@ -43,7 +63,12 @@ module Terradactyl
         File.write data_file, JSON.pretty_generate(report)
         print_ok 'Done!'
       end
+
+      def terraform_latest
+        Terradactyl::Terraform::VersionManager.latest
+      end
     end
+    # rubocop:enable Metrics/BlockLength
 
     #################################################################
     # GENERIC TASKS
@@ -247,7 +272,7 @@ module Terradactyl
       when 1
         Stacks.error!(@stack)
         print_crit "Plan failed: #{@stack.name}"
-        @stack.print_plan
+        @stack.print_error
         throw :error
       when 2
         Stacks.dirty!(@stack)
@@ -289,13 +314,7 @@ module Terradactyl
       print_ok "Cleaned: #{@stack.name}"
     end
 
-    #################################################################
-    # HIDDEN TARGETED STACK TASKS
-    # * These tasks are destructive in nature and do not require
-    # regular use.
-    #################################################################
-
-    desc 'apply NAME', 'Apply an individual stack, by name', hide: true
+    desc 'apply NAME', 'Apply an individual stack, by name'
     def apply(name)
       @stack ||= Stack.new(name)
       print_warning "Applying: #{@stack.name}"
@@ -307,7 +326,7 @@ module Terradactyl
       end
     end
 
-    desc 'refresh NAME', 'Refresh state on an individual stack, by name', hide: true
+    desc 'refresh NAME', 'Refresh state on an individual stack, by name'
     def refresh(name)
       @stack ||= Stack.new(name)
       print_crit "Refreshing: #{@stack.name}"
@@ -319,7 +338,7 @@ module Terradactyl
       end
     end
 
-    desc 'destroy NAME', 'Destroy an individual stack, by name', hide: true
+    desc 'destroy NAME', 'Destroy an individual stack, by name'
     def destroy(name)
       @stack ||= Stack.new(name)
       print_crit "Destroying: #{@stack.name}"
@@ -329,6 +348,63 @@ module Terradactyl
         Stacks.error!(@stack)
         print_crit "Failed to apply changes: #{@stack.name}"
       end
+    end
+
+    #################################################################
+    # PROJECT-LEVEL UTILITY TASKS
+    # * These tasks are managing project-wide characteristics or
+    # invoking useful commands.
+    #################################################################
+
+    desc 'install COMPONENT', 'Installs specified component'
+    long_desc <<~LONGDESC
+       The `terradactyl install COMPONENT` subcommand perfoms installations of
+      prerequisties. At present, only Terraform binaries are supported.
+       Here are a few examples:
+       # Install latest
+      `terradactyl install terraform`
+       # Install pessimistic version
+      `terradactyl install terraform --version="~> 0.13.0"`
+       # Install ranged version
+      `terradactyl install terraform --version=">= 0.14.5, <= 0.14.7"`
+       # Install explicit version
+      `terradactyl install terraform --version=0.15.0-beta2`
+
+    LONGDESC
+    option :version, type: :string, default: 'latest'
+    # rubocop:disable Metrics/AbcSize
+    def install(component)
+      case component.to_sym
+      when :terraform
+        print_warning "Installing: #{component}, version: #{options[:version]}"
+        version = options[:version] == 'latest' ? terraform_latest : options[:version]
+        Terradactyl::Terraform::VersionManager.reset!
+        Terradactyl::Terraform::VersionManager.version = version
+        Terradactyl::Terraform::VersionManager.install
+        if Terradactyl::Terraform::VersionManager.binary
+          print_ok "Installed: #{Terradactyl::Terraform::VersionManager.binary}"
+        end
+      else
+        msg = %(Operation not supported -- I don't know how to install: #{component})
+        print_crit msg
+        exit 1
+      end
+    end
+    # rubocop:enable Metrics/AbcSize
+
+    desc 'upgrade NAME', 'Upgrade an individual stack, by name'
+    def upgrade(name)
+      @stack ||= Stack.new(name)
+      print_warning "Upgrading: #{@stack.name}"
+      if @stack.upgrade.zero?
+        print_ok "Upgraded: #{@stack.name}"
+      else
+        Stacks.error!(@stack)
+        print_crit "Failed to upgrade: #{@stack.name}"
+      end
+    rescue Terradactyl::Terraform::Commands::UnsupportedCommandError => e
+      print_crit "Error: #{e.message}"
+      exit 1
     end
   end
   # rubocop:enable Metrics/ClassLength